Extract JRubyVersion into jruby_executable lib module
Move the JRubyVersion type out of the jruby_release_check binary into a
dedicated jruby_version module in the jruby_executable library and
re-export it, mirroring how shared exposes RubyDownloadVersion. The
version-only tests move with the type; binary-specific tests stay in the
release-check binary.
