+++ jruby_executable/src/jruby_version.rs
@@ -1,0 +1,129 @@
+use std::fmt;
+
+#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
+pub struct JRubyVersion {
+    major: u32,
+    minor: u32,
+    patch: u32,
+    extra: u32,
+}
+
+impl JRubyVersion {
+    pub fn parse(s: &str) -> Result<Self, String> {
+        let parts: Vec<&str> = s.split('.').collect();
+        if parts.len() != 4 {
+            return Err(format!(
+                "Invalid JRuby version '{s}': expected 4 parts (X.Y.Z.W)"
+            ));
+        }
+        let major = parts[0]
+            .parse()
+            .map_err(|_| format!("Invalid major version in '{s}'"))?;
+        let minor = parts[1]
+            .parse()
+            .map_err(|_| format!("Invalid minor version in '{s}'"))?;
+        let patch = parts[2]
+            .parse()
+            .map_err(|_| format!("Invalid patch version in '{s}'"))?;
+        let extra = parts[3]
+            .parse()
+            .map_err(|_| format!("Invalid extra version in '{s}'"))?;
+        Ok(Self {
+            major,
+            minor,
+            patch,
+            extra,
+        })
+    }
+}
+
+impl fmt::Display for JRubyVersion {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(
+            f,
+            "{}.{}.{}.{}",
+            self.major, self.minor, self.patch, self.extra
+        )
+    }
+}
+
+impl serde::Serialize for JRubyVersion {
+    fn serialize<S: serde::Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
+        serializer.serialize_str(&self.to_string())
+    }
+}
+
+impl std::str::FromStr for JRubyVersion {
+    type Err = String;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        Self::parse(s)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_parse_valid_version() {
+        let v = JRubyVersion::parse("9.4.7.0").unwrap();
+        assert_eq!(v.major, 9);
+        assert_eq!(v.minor, 4);
+        assert_eq!(v.patch, 7);
+        assert_eq!(v.extra, 0);
+    }
+
+    #[test]
+    fn test_parse_invalid_versions() {
+        assert!(JRubyVersion::parse("9.4.7").is_err());
+        assert!(JRubyVersion::parse("9.4").is_err());
+        assert!(JRubyVersion::parse("9").is_err());
+        assert!(JRubyVersion::parse("").is_err());
+        assert!(JRubyVersion::parse("9.4.7.0.1").is_err());
+        assert!(JRubyVersion::parse("a.b.c.d").is_err());
+    }
+
+    #[test]
+    fn test_version_display() {
+        let v = JRubyVersion::parse("9.4.7.0").unwrap();
+        assert_eq!(v.to_string(), "9.4.7.0");
+
+        let v = JRubyVersion::parse("10.1.0.0").unwrap();
+        assert_eq!(v.to_string(), "10.1.0.0");
+    }
+
+    #[test]
+    fn test_version_ordering() {
+        let v940 = JRubyVersion::parse("9.4.0.0").unwrap();
+        let v947 = JRubyVersion::parse("9.4.7.0").unwrap();
+        let v9415 = JRubyVersion::parse("9.4.15.0").unwrap();
+        let v1010 = JRubyVersion::parse("10.1.0.0").unwrap();
+
+        assert!(v947 > v940);
+        assert!(v9415 > v947);
+        assert!(v1010 > v9415);
+        assert!(v940 < v947);
+        assert_eq!(v947, JRubyVersion::parse("9.4.7.0").unwrap());
+    }
+
+    #[test]
+    fn test_version_from_str() {
+        let v: JRubyVersion = "9.4.7.0".parse().unwrap();
+        assert_eq!(v.to_string(), "9.4.7.0");
+    }
+
+    #[test]
+    fn test_version_serialize_json() {
+        let v = JRubyVersion::parse("9.4.7.0").unwrap();
+        let json = serde_json::to_string(&v).unwrap();
+        assert_eq!(json, "\"9.4.7.0\"");
+
+        let versions = vec![
+            JRubyVersion::parse("10.1.0.0").unwrap(),
+            JRubyVersion::parse("9.4.15.0").unwrap(),
+        ];
+        let json = serde_json::to_string_pretty(&versions).unwrap();
+        assert_eq!(json, "[\n  \"10.1.0.0\",\n  \"9.4.15.0\"\n]");
+    }
+}
